+++ lib/Completion/Bridge/TolerantParser/WorseReflection/WorseNamedParameterCompletor.php
@@ -12,6 +12,7 @@ use Microsoft\PhpParser\Node\Expression\ObjectCreationExpression;
 use Microsoft\PhpParser\Node\Expression\ScopedPropertyAccessExpression;
 use Microsoft\PhpParser\Node\Expression\Variable;
 use Microsoft\PhpParser\Node\QualifiedName;
+use Microsoft\PhpParser\Node\StringLiteral;
 use Phpactor\Completion\Bridge\TolerantParser\TolerantCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\Helper\NodeQuery;
 use Phpactor\Completion\Core\Formatter\ObjectFormatter;
@@ -35,7 +36,7 @@ class WorseNamedParameterCompletor implements TolerantCompletor
 
     public function complete(Node $node, TextDocument $source, ByteOffset $offset): Generator
     {
-        if (null === $creation = NodeQuery::firstAncestorOrSelfInVia(
+        $subject = NodeQuery::firstAncestorOrSelfInVia(
             $node,
             [
                 MemberAccessExpression::class,
@@ -46,20 +47,22 @@ class WorseNamedParameterCompletor implements TolerantCompletor
                 ArgumentExpression::class,
                 ArgumentExpressionList::class
             ]
-        )) {
+        );
+
+        if (null === $subject) {
             return true;
         }
 
-        if ($node instanceof Variable) {
+        if ($node instanceof Variable || $node instanceof StringLiteral) {
             return true;
         }
 
-        if ($creation instanceof ObjectCreationExpression) {
-            return yield from $this->fromObjectCreation($creation);
+        if ($subject instanceof ObjectCreationExpression) {
+            return yield from $this->fromObjectCreation($subject);
         }
 
-        if ($creation instanceof CallExpression) {
-            return yield from $this->fromCallExpression($creation);
+        if ($subject instanceof CallExpression) {
+            return yield from $this->fromCallExpression($subject);
         }
 
         return true;