+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Improvements:
 
 Bug fixes:
 
+  - Do not show named parameters after string literal argument #2259
   - Fix "instanceof" behavior for statically reflected classes #2273
   - Fix behavior when user cancels type selection on goto type #2270
   - Fix docblock parsing of `array<'quoted'|'strings'>` #2264